+++ android/src/main/java/ly/count/android/sdk/react/CountlyReactNative.java
@@ -74,7 +74,7 @@ class CountlyReactException extends Exception {
 public class CountlyReactNative extends ReactContextBaseJavaModule implements LifecycleEventListener {
 
     public static final String TAG = "CountlyRNPlugin";
-    private String COUNTLY_RN_SDK_VERSION_STRING = "20.11.10";
+    private String COUNTLY_RN_SDK_VERSION_STRING = "20.11.11";
     private String COUNTLY_RN_SDK_NAME = "js-rnb-android";
 
     private static CountlyConfig config = new CountlyConfig();
@@ -130,8 +130,8 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
         this.config.setServerURL(serverUrl);
         this.config.setAppKey(appKey);
 
-          Countly.sharedInstance().COUNTLY_SDK_NAME = COUNTLY_RN_SDK_NAME;
-          Countly.sharedInstance().COUNTLY_SDK_VERSION_STRING = COUNTLY_RN_SDK_VERSION_STRING;
+        Countly.sharedInstance().COUNTLY_SDK_NAME = COUNTLY_RN_SDK_NAME;
+        Countly.sharedInstance().COUNTLY_SDK_VERSION_STRING = COUNTLY_RN_SDK_VERSION_STRING;
 
         this.config.setContext(_reactContext);
         Activity activity = getActivity();
@@ -181,7 +181,7 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
         Boolean result = Countly.sharedInstance().hasBeenCalledOnStart();
         promise.resolve(result);
     }
-    
+
     @ReactMethod
     public void getCurrentDeviceId(Promise promise){
         String deviceID = Countly.sharedInstance().getDeviceID();
@@ -455,17 +455,17 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
     @ReactMethod
     public void setUserData(ReadableArray args){
         Countly.sharedInstance();
-        Map<String, String> bundle = new HashMap<String, String>();
-        bundle.put("name", args.getString(0));
-        bundle.put("username", args.getString(1));
-        bundle.put("email", args.getString(2));
-        bundle.put("organization", args.getString(3));
-        bundle.put("phone", args.getString(4));
-        bundle.put("picture", args.getString(5));
-        bundle.put("picturePath", args.getString(6));
-        bundle.put("gender", args.getString(7));
-        bundle.put("byear", String.valueOf(args.getInt(8)));
-        Countly.userData.setUserData(bundle);
+        ReadableMap userData = args.getMap(0);
+        Map<String, Object> userDataObjectMap = userData.toHashMap();
+        Map<String,String> userDataMap =new HashMap<String,String>();
+        for (Map.Entry<String, Object> entry : userDataObjectMap.entrySet()) {
+            Object value = entry.getValue();
+            if(value instanceof String){
+                userDataMap.put(entry.getKey(), (String) value);
+            }
+        }
+
+        Countly.userData.setUserData(userDataMap);
         Countly.userData.save();
     }
 
@@ -481,7 +481,7 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
         this.channelName = args.getString(1);
         this.channelDescription = args.getString(2);
         log("pushTokenType [" + messagingMode + "][" + this.channelName + "][" + this.channelDescription + "]", LogLevel.INFO);
-        
+
         if (messagingMode == 0) {
             this.messagingMode = Countly.CountlyMessagingMode.PRODUCTION;
         } else {
@@ -494,7 +494,7 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
         JSONObject json = new JSONObject(notification);
         String notificationString = json.toString();
         log("onNotification [" + notificationString + "]", LogLevel.INFO);
-        
+
         if(notificationListener != null){
             //there is a listener for notifications, send the just received notification to it
             log("onNotification, listener exists", LogLevel.INFO);
@@ -546,19 +546,32 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
         }
         CountlyPush.useAdditionalIntentRedirectionChecks = true;
         CountlyPush.init(activity.getApplication(), messagingMode);
-        FirebaseApp.initializeApp(context);
-        FirebaseInstanceId.getInstance().getInstanceId()
-                .addOnCompleteListener(new OnCompleteListener<InstanceIdResult>() {
-                    @Override
-                    public void onComplete(Task<InstanceIdResult> task) {
-                        if (!task.isSuccessful()) {
-                            log("getInstanceId failed", task.getException(), LogLevel.WARNING);
-                            return;
-                        }
-                        String token = task.getResult().getToken();
-                        CountlyPush.onTokenRefresh(token);
+        try{
+            FirebaseApp.initializeApp(context);
+            FirebaseInstanceId firebaseInstanceId = FirebaseInstanceId.getInstance();
+            if(firebaseInstanceId == null) {
+                log("askForNotificationPermission, firebaseInstanceId is null", LogLevel.WARNING);
+                return;
+            }
+            Task<InstanceIdResult> instanceIdResultTask = firebaseInstanceId.getInstanceId();
+            if(instanceIdResultTask == null) {
+                log("askForNotificationPermission, instanceIdResultTask is null", LogLevel.WARNING);
+                return;
+            }
+            instanceIdResultTask.addOnCompleteListener(new OnCompleteListener<InstanceIdResult>() {
+                @Override
+                public void onComplete(Task<InstanceIdResult> task) {
+                    if (!task.isSuccessful()) {
+                        log("askForNotificationPermission, getInstanceId failed", task.getException(), LogLevel.WARNING);
+                        return;
                     }
-                });
+                    String token = task.getResult().getToken();
+                    CountlyPush.onTokenRefresh(token);
+                }
+            });
+        }catch(Exception exception){
+            log("askForNotificationPermission, Firebase exception",exception, LogLevel.WARNING);
+        }
     }
 
     @ReactMethod
@@ -872,7 +885,7 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
 
     @ReactMethod
     public void getFeedbackWidgets(final Promise promise)
-     {
+    {
         Countly.sharedInstance().feedback().getAvailableFeedbackWidgets(new RetrieveFeedbackWidgets() {
             @Override
             public void onFinished(List<CountlyFeedbackWidget> retrievedWidgets, String error) {
@@ -891,11 +904,11 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
                 promise.resolve(retrievedWidgetsArray);
             }
         });
-    } 
+    }
 
     @ReactMethod
     public void getAvailableFeedbackWidgets(final Promise promise)
-     {
+    {
         Countly.sharedInstance().feedback().getAvailableFeedbackWidgets(new RetrieveFeedbackWidgets() {
             @Override
             public void onFinished(List<CountlyFeedbackWidget> retrievedWidgets, String error) {
@@ -910,7 +923,7 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
                 promise.resolve(retrievedWidgetsMap);
             }
         });
-    } 
+    }
 
     @ReactMethod
     public void presentFeedbackWidget(ReadableArray args, final Promise promise) {
@@ -940,18 +953,18 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
                 }
             }
         });
-    } 
-    
+    }
+
 
     @ReactMethod
     public void replaceAllAppKeysInQueueWithCurrentAppKey() {
         Countly.sharedInstance().requestQueueOverwriteAppKeys();
-    } 
-    
+    }
+
     @ReactMethod
     public void removeDifferentAppKeysFromQueue() {
         Countly.sharedInstance().requestQueueEraseAppKeysRequests();
-    } 
+    }
 
     @ReactMethod
     public void setEventSendThreshold(ReadableArray args){
@@ -1041,7 +1054,7 @@ public class CountlyReactNative extends ReactContextBaseJavaModule implements Li
         }
         this.config.setMetricOverride(customMetric);
     }
-    
+
     enum LogLevel {INFO, DEBUG, VERBOSE, WARNING, ERROR}
     static void log(String message, LogLevel logLevel)  {
         log(message, null, logLevel);